requirements.txt
Django==1.5
South==0.7.6
dj-database-url==0.2.1
django-social-auth==0.7.22
django-tastypie==0.9.14
facepy==0.8.4
gunicorn==0.17.2
httplib2==0.8
mimeparse==0.1.3
oauth2==1.5.211
pep8==1.4.5
psycopg2==2.4.6
python-dateutil==2.1
python-openid==2.2.5
requests==1.1.0
selenium==2.31.0
six==1.3.0
wsgiref==0.1.2
